backend validation

+++ backend/controller/user/userAddressController.js
@@ -1,29 +1,29 @@
+import mongoose from 'mongoose'
 import { Address } from '../../model/address.js'
 import { User } from '../../model/user.js'
 
 const addAddress = async (req, res) => {
     try {
-        const { firstName, lastName, buildingName, street, city, district, state, pincode, phoneNumber } = req.body.address
+        const { firstName = '', lastName = '', buildingName = '', street = '', city = '', state = '', pincode, phoneNumber } = req.body
+        console.log(req.body)
         const existingAddress = await Address.findOne({
-            firstName,
-            lastName,
-            buildingName,
-            street,
-            district,
-            city,
-            state,
+            firstName: firstName.toLowerCase(),
+            lastName: lastName.toLowerCase(),
+            buildingName: buildingName.toLowerCase(),
+            street: street.toLowerCase(),
+            city: city.toLowerCase(),
+            state: state.toLowerCase(),
             pincode,
             phoneNumber
         })
         if (existingAddress) return res.status(409).json({ message: 'Address already exists' })
         const address = await Address.create({
-            firstName,
-            lastName,
-            buildingName,
-            street,
-            district,
-            city,
-            state,
+            firstName: firstName.toLowerCase(),
+            lastName: lastName.toLowerCase(),
+            buildingName: buildingName.toLowerCase(),
+            street: street.toLowerCase(),
+            city: city.toLowerCase(),
+            state: state.toLowerCase(),
             pincode,
             phoneNumber
         })
@@ -31,11 +31,11 @@ const addAddress = async (req, res) => {
         if (user && address) return res.status(201).json({ message: 'Address added sucessfully' })
     } catch (e) {
         if (e.name === 'ValidationError') {
-            const errMsg = []
+            const message = []
             for (let error in e.errors) {
-                errMsg.push(e.errors[error].properties.message)
+                message.push(e.errors[error].properties.message)
             }
-            return res.status(400).json({ errMsg })
+            return res.status(400).json({ message })
         }
         return res.status(500).json({ message: 'Failed to add address' })
     }
@@ -44,26 +44,25 @@ const addAddress = async (req, res) => {
 const updateAddress = async (req, res) => {
     const { id } = req.params
     try {
-        const { firstName, lastName, buildingName, street, city, district, state, pincode, phoneNumber } = req.body.address
+        const { firstName = '', lastName = '', buildingName = '', street = '', city = '', state = '', pincode, phoneNumber } = req.body
         const address = await Address.findByIdAndUpdate(id, {
-            firstName,
-            lastName,
-            buildingName,
-            street,
-            district,
-            city,
-            state,
+            firstName: firstName.toLowerCase(),
+            lastName: lastName.toLowerCase(),
+            buildingName: buildingName.toLowerCase(),
+            street: street.toLowerCase(),
+            city: city.toLowerCase(),
+            state: state.toLowerCase(),
             pincode,
             phoneNumber
         }, { new: true, runValidators: true })
         if (address) return res.status(200).json({ message: 'Address updated sucessfully' })
     } catch (e) {
         if (e.name === 'ValidationError') {
-            const errMsg = []
+            const message = []
             for (let error in e.errors) {
-                errMsg.push(e.errors[error].properties.message)
+                message.push(e.errors[error].properties.message)
             }
-            return res.status(400).json({ errMsg })
+            return res.status(400).json({ errMsg: message })
         }
         return res.status(500).json({ message: 'Failed to update address' })
     }
@@ -71,6 +70,11 @@ const updateAddress = async (req, res) => {
 
 const deleteAddress = async (req, res) => {
     const { id } = req.params
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(400).json({ message: 'Invalid address ID' })
+    }
+    const address = await Address.findById(id)
+    if (!address) return res.status(400).json({ message: 'Address with ID not found' })
     try {
         const response = await Address.findByIdAndDelete(id)
         const user = await User.findByIdAndUpdate(req.userId, {
@@ -78,7 +82,6 @@ const deleteAddress = async (req, res) => {
         }, { new: true })
         if (response && user) return res.status(200).json({ message: 'Address deleted sucessfully' })
     } catch (e) {
-        console.log(e)
         return res.status(500).json({ message: 'Failed to delete address' })
     }
 }
